requirements.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.2
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mypy==1.8.0
mypy-extensions==1.0.0
packaging==23.2
pluggy==1.4.0
python-dotenv==1.0.1
typing_extensions==4.10.0
Werkzeug==3.0.1
